# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SHOPIN",
        "description": "SHOPIN API's integration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/shopiners": {
            "get": {
                "tags": [
                    "Shopiners"
                ],
                "summary": "List paginated shopiners",
                "description": "Returns a paginated list of shopiners with full avatar URLs",
                "operationId": "97c59d1b43dcbbf1a6119d9d7992a7bd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Search by username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter users with average rating greater than or equal to this value",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of shopiners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 11
                                                            },
                                                            "firstname": {
                                                                "type": "string",
                                                                "example": "John"
                                                            },
                                                            "lastname": {
                                                                "type": "string",
                                                                "example": "Doe"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "johndoe"
                                                            },
                                                            "voyage_mode": {
                                                                "type": "string",
                                                                "example": "Fast"
                                                            },
                                                            "avatar": {
                                                                "type": "string",
                                                                "example": "http://127.0.0.1:8000/storage/uploads/avatars/avatar1.png"
                                                            },
                                                            "photo_verified_at": {
                                                                "type": "string",
                                                                "example": "2025-01-10 12:00:00",
                                                                "nullable": true
                                                            },
                                                            "average_rating": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 4.5
                                                            },
                                                            "total_posts": {
                                                                "type": "integer",
                                                                "example": 23
                                                            },
                                                            "total_reviews": {
                                                                "type": "integer",
                                                                "example": 10
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shopiner/profile/{id}": {
            "get": {
                "tags": [
                    "Shopiners"
                ],
                "summary": "Get Shopiner profile",
                "description": "Returns a Shopiner's profile including average rating, total posts, total reviews, total views, and posts created by him",
                "operationId": "56ae8f022ab65b53815973e8248f700b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Shopiner ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shopiner profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "shopiner": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 11
                                                        },
                                                        "firstname": {
                                                            "type": "string",
                                                            "example": "Malak"
                                                        },
                                                        "lastname": {
                                                            "type": "string",
                                                            "example": "Salame"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "Louka"
                                                        },
                                                        "voyage_mode": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "avatar": {
                                                            "type": "string",
                                                            "example": "http://127.0.0.1:8000/storage/avatar.png"
                                                        },
                                                        "photo_verified_at": {
                                                            "type": "string",
                                                            "example": "2024-10-02 23:11:45",
                                                            "nullable": true
                                                        },
                                                        "average_rating": {
                                                            "type": "string",
                                                            "example": "3.0000"
                                                        },
                                                        "total_posts": {
                                                            "type": "integer",
                                                            "example": 18
                                                        },
                                                        "total_reviews": {
                                                            "type": "integer",
                                                            "example": 18
                                                        },
                                                        "total_views": {
                                                            "type": "integer",
                                                            "example": 120
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "posts": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 30
                                                            },
                                                            "titre": {
                                                                "type": "string",
                                                                "example": "Manteau old style"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Manteau camel Mango"
                                                            },
                                                            "photos": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "http://127.0.0.1:8000/storage/uploads/posts/TqhzM91f7GYgu5V6NwiYYNi1hKZEk9TlA26dZJ8a.webp"
                                                                }
                                                            },
                                                            "prix": {
                                                                "type": "string",
                                                                "example": "300.00"
                                                            },
                                                            "etat": {
                                                                "type": "string",
                                                                "example": "Bon état"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Shopiner not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shopiner not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}